Move X main loop and event handling to own file
The structure of the code is kept unchanged. The next step is to make
all the event handlers member functions of XMainLoop.
