+++ src/hlwmcommon.h
@@ -1,7 +1,9 @@
 #pragma once
 
 #include <X11/X.h>
+#include <tuple>
 #include <unordered_map>
+#include <vector>
 
 class Root;
 class Client;
@@ -15,6 +17,7 @@ public:
     //! The Client object for a window or nullptr if unmanaged.
     Client* client(Window window);
     const std::unordered_map<Window, Client*>& clients();
+    static std::pair<int,std::string> callCommand(const std::vector<std::string>& call);
 private:
     Root* root_;
 };
